Fixed header link padding; Added gallery buttons
Fixed the size of the padding around the links in the header to make them easier to click. Fixed issues with the dropdown menus in the header. Hopefully fixed gallery preset filter bug where it wouldn't apply when first loading page. Added buttons to go to the next or previous picture in gallery. Still need to make that responsive.